widgets/store [nfc]: Move data-managing widgets to their own file
