+++ lib/widgets/content.dart
@@ -4,7 +4,7 @@ import 'package:html/dom.dart' as dom;
 import '../api/model/model.dart';
 import '../model/content.dart';
 import '../model/store.dart';
-import 'app.dart';
+import 'store.dart';
 
 /// The font size for message content in a plain unstyled paragraph.
 const double kBaseFontSize = 14;
